Fix #2242: Move DummyCookieJar into cookiejar.py
