+++ aiohttp/__init__.py
@@ -10,7 +10,7 @@ from .http import (HttpVersion, HttpVersion10, HttpVersion11,  # noqa
                    WSMsgType, WSCloseCode, WSMessage, WebSocketError)  # noqa
 from .streams import *  # noqa
 from .multipart import *  # noqa
-from .cookiejar import CookieJar  # noqa
+from .cookiejar import *  # noqa
 from .payload import *  # noqa
 from .payload_streamer import *  # noqa
 from .resolver import *  # noqa
@@ -23,6 +23,7 @@ except ImportError:
 
 
 __all__ = (client.__all__ +  # noqa
+           cookiejar.__all__ +  # noqa
            formdata.__all__ +  # noqa
            helpers.__all__ +  # noqa
            multipart.__all__ +  # noqa
@@ -31,5 +32,5 @@ __all__ = (client.__all__ +  # noqa
            streams.__all__ +  # noqa
            ('hdrs', 'HttpVersion', 'HttpVersion10', 'HttpVersion11',
             'WSMsgType', 'WSCloseCode',
-            'WebSocketError', 'WSMessage', 'CookieJar',
+            'WebSocketError', 'WSMessage',
            ) + workers)
