+++ changes/2242.feature
@@ -1,0 +1,1 @@
+Move DummyCookieJar into cookiejar.py
